requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
docs==0.1.0
docx==0.2.4
Flask==2.0.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
lxml==4.7.1
MarkupSafe==2.0.1
numpy==1.22.0
pandas==1.3.5
Pillow==9.0.0
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2021.3
requests==2.27.0
six==1.16.0
urllib3==1.26.7
Werkzeug==2.0.2
